requirements.txt
Flask
flask-cors
PyPDF2
python-dotenv
pandas
Pillow
google-generativeai==0.7.0